requirements.txt
# Database
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2

# CLI utilities
click==8.1.7
tabulate==0.9.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

